refine print_op_decode()
It now decodes simple-ish inout parameters as well as in-parameters, and
produces helper variables for out-parameters. These are passed correctly by
reference.

The return value for NegativeReturn exceptions is now masked down to 16 or 8
bits long, unsigned, as is the exception content. Since exception labels
cannot be specified in this version of libIDL, NegativeReturn exceptions are
always given the MSG_ERROR label which is #included from <kernel/message.h>.