recognize the PagerTarget attribute for operations
This means that IPC is always sent to the return value of L4_Pager(), which is
never L4_nilthread (except for sigma0, which is a L4.X2 special case and thus
outside µIDL's scope). The non-autotest case in tests/attr-pagertarget.idl
provides inputs for the four cases, i.e. void and non-void functions, with and
without PagerTarget.