use memcpy(3) on stub receive string buffers where appropriate
String items can only be received in a single buffer. So for each string item,
the stub generator selects the first among [reply, exn0_0, exn1_0, ...] and
receives the string item there. Depending on what was received, this may be
the wrong address. memcpy(3) is used to copy data to its correct position.