remove the muidl::mapping native type
It seems to have been intended for passing non-aligned memory between IPC
participants as a big sequence of MapItems. This has two things going against
it: first, since it was broken it can be assumed that it was also entirely
untested; and second, no need was demonstrated for such a mechanism.

A minor third issue was the fuck-awful mapping encode/decode LLVM shit in
mapping.c, mercifully removed.