remove GHashTable use in except.c, struct.c, util.c
They're replaced with strset and strmap as applicable. As for efficiency: this
is application code, so a few tree-walks don't matter if the interface is
nice. This is, and it removes lines, so in it goes.

The new warn_once() isn't actually tested as nothing is generating a warning.
Presumed correct regardless.